Refactoring kafka collectors into common file
